Terms page little fix (#1229)
* refactored two wiki created files

* refactored two wiki created files

* update codecimate yml file

* update codecimate yml file

* refactoring pages and prop types

* imports update

* reducing pages to meet method line of 200

* reducing pages to meet method line of 200

* updating the code climate file

* wiki created action bar and footer

* wiki created action bar and footer

* exported types

* table column refactor

* table col date refactor

* table col date refactor

* refacoting functions

* refactoring functions in WikiInsight Page

* Line

* fixed email issue
