+++ .codeclimate.yml
@@ -24,7 +24,7 @@ checks:
   method-lines:
     enabled: true
     config:
-      threshold: 25
+      threshold: 200
   nested-control-flow:
     enabled: true
     config:
