+++ src/components/Admin/WikiCreatedInsight/PromoteCreatedWikisModal.tsx
@@ -4,63 +4,38 @@ import {
   ModalOverlay,
   ModalContent,
   ModalBody,
-  Box,
-  Text,
-  HStack,
   Icon,
-  Button,
   Flex,
-  Heading,
-  VStack,
-  AspectRatio,
-  Link,
-  Stack,
-  Select,
   useToast,
 } from '@chakra-ui/react'
-import { Step, Steps, useSteps } from 'chakra-ui-steps'
+import { useSteps } from 'chakra-ui-steps'
 import { RiCloseLine } from 'react-icons/ri'
 import {
   useGetSearchedWikisByTitleQuery,
   usePostPromotedWikiMutation,
   useGetAllPromotedWikiCountQuery,
 } from '@/services/admin'
-import DisplayAvatar from '@/components/Elements/Avatar/DisplayAvatar'
-import { WikiImage } from '@/components/WikiImage'
-import { shortenAccount } from '@/utils/textUtils'
-import { getReadableDate } from '@/utils/DataTransform/getFormattedDate'
-import { Image } from '@everipedia/iq-utils'
-import { getWikiImageUrl } from '@/utils/WikiUtils/getWikiImageUrl'
+import { PromoteCreatedWikisModalProps } from '@/types/admin'
+import { getWikiIdUsingLevel } from '@/utils/AdminUtils/dataUpdate'
+import { PromoteModalContent } from './PromotedWikiContent'
 
-export const PromoteCreatedWikisModal = ({
-  onClose = () => {},
-  isOpen = false,
-  wikiChosenTitle,
-  wikiChosenId,
-  hideFunc,
-  ...rest
-}: {
-  isOpen: boolean
-  onClose: () => void
-  wikiChosenTitle: string
-  wikiChosenId: string
-  hideFunc: () => void
-}) => {
+export const PromoteCreatedWikisModal = (
+  props: PromoteCreatedWikisModalProps,
+) => {
+  const {
+    onClose = () => {},
+    isOpen = false,
+    wikiChosenTitle,
+    wikiChosenId,
+    hideFunc,
+    ...rest
+  } = props
   const [step2Titles, setStep2Titles] = useState('Promote to Homepage')
   const [buttonOne, setbuttonOne] = useState('Promote to Hero section')
   const [buttonTwo, setbuttonTwo] = useState('Promote to Trending wikis')
   const [initGetSearchedWikis, setInitGetSearchedWikis] =
     useState<boolean>(true)
   const { data: promotedWikis } = useGetAllPromotedWikiCountQuery(0)
-
-  const getWikiIdUsingLevel = (level: number) => {
-    const wiki = promotedWikis?.filter(item => {
-      return item.promoted === level
-    })[0]
-
-    return wiki?.id
-  }
-
   const arrs = () => {
     const arr = []
     const data = promotedWikis || []
@@ -73,7 +48,6 @@ export const PromoteCreatedWikisModal = ({
     skip: initGetSearchedWikis,
   })
   const [value, setValue] = useState('2')
-
   const toast = useToast()
   const ModalData = wiki?.filter(
     item => item.id === wikiChosenId && item.title === wikiChosenTitle,
@@ -88,7 +62,6 @@ export const PromoteCreatedWikisModal = ({
     { label: 'Step 2', description: step2Titles },
     { label: 'Step 3', description: 'Promotion confirmation' },
   ]
-
   const [promoteWiki, { error: posTPromoteWikiError }] =
     usePostPromotedWikiMutation()
 
@@ -127,126 +100,6 @@ export const PromoteCreatedWikisModal = ({
       isClosable: true,
     })
   }
-  const getWiki = (
-    <>
-      {Data && (
-        <HStack
-          bgColor="cardBg"
-          justifyContent="flex-start"
-          borderWidth="2px"
-          borderColor="cardBorder"
-          borderRadius="lg"
-          borderStyle="dotted"
-          px={{ base: 3, lg: 5 }}
-          py={{ base: 3, lg: 3 }}
-          w="full"
-          align="normal"
-        >
-          <AspectRatio w={{ base: '100px', md: '140px', lg: '156px' }}>
-            <WikiImage
-              cursor="pointer"
-              flexShrink={0}
-              imageURL={getWikiImageUrl(Data.images as Image[])}
-              borderRadius="lg"
-              overflow="hidden"
-              alt="wiki"
-            />
-          </AspectRatio>
-          <Flex
-            w="90%"
-            flexDir="column"
-            justify="space-between"
-            mx="auto"
-            px={4}
-          >
-            <Flex justifyContent="space-between" mb={{ base: 0, md: 2 }}>
-              <HStack w={{ base: '83%', md: '70%' }}>
-                <Heading
-                  cursor="pointer"
-                  as="h2"
-                  fontSize={{ base: '16px', md: '20px' }}
-                  letterSpacing="wide"
-                  overflow="hidden"
-                  whiteSpace="nowrap"
-                  textOverflow="ellipsis"
-                  tabIndex={0}
-                  role="link"
-                >
-                  {Data.title}
-                </Heading>
-              </HStack>
-              {Data.categories.length ? (
-                <HStack>
-                  {Data.categories?.map((category, i) => (
-                    <Link
-                      key={i}
-                      href={`/categories/${category.id}`}
-                      display={{ base: 'none', md: 'block' }}
-                      color="brand.500"
-                      fontWeight="bold"
-                      cursor="pointer"
-                      fontSize={{ base: '10px', lg: '12px' }}
-                    >
-                      {category.title ? category.title : ''}
-                    </Link>
-                  ))}
-                </HStack>
-              ) : (
-                <Text> </Text>
-              )}
-            </Flex>
-            <Box
-              mb="2"
-              mt="-2%"
-              maxW={{ base: '70%', lg: '80%' }}
-              overflow="hidden"
-            >
-              <Text fontSize="sm" display={{ base: 'none', md: 'flex' }}>
-                {Data.summary ? Data.summary : 'No Summary'}
-              </Text>
-            </Box>
-            <Stack
-              direction={{ base: 'column', md: 'row' }}
-              justifyContent="space-between"
-              w="full"
-            >
-              <Box>
-                <HStack flex="1">
-                  <DisplayAvatar
-                    address={Data.author?.id}
-                    avatarIPFS={Data.author.profile?.avatar}
-                    size={20}
-                    alt="unknown"
-                  />
-                  <Text fontSize="14px" color="linkColor">
-                    <Link
-                      href={`/account/${Data.author?.id}`}
-                      color="brand.500"
-                      fontWeight="bold"
-                    >
-                      {/* eslint-disable no-nested-ternary */}
-                      {Data.author?.profile?.username
-                        ? Data.author.profile.username
-                        : Data.author?.id
-                        ? shortenAccount(Data.author.id)
-                        : 'Unknown'}
-                    </Link>
-                  </Text>
-                </HStack>
-              </Box>
-              <Box>
-                {Data.updated && (
-                  <Text mt="1" fontSize="sm" opacity={0.6} whiteSpace="nowrap">
-                    {getReadableDate(Data.updated)}
-                  </Text>
-                )}
-              </Box>
-            </Stack>
-          </Flex>
-        </HStack>
-      )}
-    </>
-  )
 
   const TrendingwikiSelected = async () => {
     if (activeStep === 0) {
@@ -267,8 +120,7 @@ export const PromoteCreatedWikisModal = ({
           level: Number(value),
         })
         handlePromoteWiki({ id: wikiChosenId, level: Number(value) })
-
-        const id = getWikiIdUsingLevel(+value)
+        const id = getWikiIdUsingLevel(+value, promotedWikis)
         if (id) {
           handlePromoteWiki({ id, level: 0 })
         }
@@ -276,7 +128,7 @@ export const PromoteCreatedWikisModal = ({
         Close()
       } else {
         handlePromoteWiki({ id: wikiChosenId, level: 1 })
-        const id = getWikiIdUsingLevel(1)
+        const id = getWikiIdUsingLevel(1, promotedWikis)
         if (id) {
           handlePromoteWiki({ id, level: 0 })
         }
@@ -301,65 +153,6 @@ export const PromoteCreatedWikisModal = ({
     }
   }
   if (!isOpen) return null
-
-  const content = (
-    <>
-      {activeStep === 0 && (
-        <Text textAlign="center">
-          Select the appropriate action you would like to take for this wiki
-        </Text>
-      )}
-      {activeStep === 1 && (
-        <>
-          <VStack gap={4}>
-            {step2Titles === 'Promote to Trending wiki' && (
-              <Box w="full">
-                <Text fontWeight="bold" py="1">
-                  Select slot
-                </Text>
-                <Select
-                  cursor="pointer"
-                  onChange={e => setValue(e.target.value)}
-                  defaultValue={promotedWikis?.length}
-                >
-                  {promotedWikis &&
-                    [...promotedWikis]
-                      ?.sort((a, b) => a.promoted - b.promoted)
-                      ?.slice(1)
-                      ?.map(item => (
-                        <option value={item.promoted}>
-                          SLOT {item.promoted - 1} - {item.title}
-                        </option>
-                      ))}
-                  {promotedWikis && (
-                    <option value={promotedWikis && +promotedWikis.length + 1}>
-                      New Slot
-                    </option>
-                  )}
-                </Select>
-              </Box>
-            )}
-            {getWiki}
-          </VStack>
-        </>
-      )}
-      {activeStep === 2 && (
-        <>
-          {step2Titles === 'Promote to Trending wiki' ? (
-            <Text textAlign="center">
-              You are about to promote a wiki to the Trending wiki section of
-              the homepage. Do you wish to continue this action?
-            </Text>
-          ) : (
-            <Text textAlign="center">
-              You are about to promote a wiki to the hero section of the
-              homepage. Do you wish to continue this action?
-            </Text>
-          )}
-        </>
-      )}
-    </>
-  )
 
   return (
     <Modal
@@ -386,58 +179,18 @@ export const PromoteCreatedWikisModal = ({
               alignSelf="center"
             />
           </Flex>
-
-          <VStack px="5" py="3" gap={4}>
-            <Text fontSize="xl" textAlign="start" w="100%" fontWeight="bold">
-              Promote
-            </Text>
-            <Flex flexDir="column" width="100%">
-              <Steps
-                labelOrientation="vertical"
-                colorScheme="brand"
-                activeStep={activeStep}
-                size="sm"
-              >
-                {steps.map(({ label, description }) => (
-                  <Step
-                    textAlign="left"
-                    label={label}
-                    key={label}
-                    description={description}
-                  >
-                    <Box pt="10" pb="10">
-                      {content}
-                    </Box>
-                  </Step>
-                ))}
-              </Steps>
-              <Flex width="100%" justify="center" pt={4} pb={5}>
-                <HStack gap={3}>
-                  <Button
-                    p={4}
-                    onClick={HompageSelected}
-                    size="sm"
-                    variant="ghost"
-                    fontSize="xs"
-                  >
-                    {buttonOne}
-                  </Button>
-                  <Button
-                    size="sm"
-                    fontSize="xs"
-                    borderWidth="1px"
-                    onClick={TrendingwikiSelected}
-                    disabled={
-                      !promotedWikis?.length &&
-                      buttonTwo === 'Promote to Trending wikis'
-                    }
-                  >
-                    {buttonTwo}
-                  </Button>
-                </HStack>
-              </Flex>
-            </Flex>
-          </VStack>
+          <PromoteModalContent
+            activeStep={activeStep}
+            steps={steps}
+            HompageSelected={HompageSelected}
+            buttonOne={buttonOne}
+            buttonTwo={buttonTwo}
+            step2Titles={step2Titles}
+            promotedWikis={promotedWikis}
+            Data={Data}
+            setValue={setValue}
+            TrendingwikiSelected={TrendingwikiSelected}
+          />
         </ModalBody>
       </ModalContent>
     </Modal>
