+++ src/types/admin.ts
@@ -1,3 +1,5 @@
+import { ChangeEvent } from 'react'
+
 export type EditorsType = {
   id: string
   title: string
@@ -175,4 +177,83 @@ export interface EditorsTable {
   latestActivity: string
   editorAddress: string
   active: boolean
+}
+
+export type InsightTableWikiEditorsProps = {
+  wikiInsightData: Editors[] | undefined
+  toggleUserFunc?: (active: boolean, id: string) => void
+  editorsIsFetching: boolean
+  hiddenEditorsIsFetching: boolean
+}
+
+export type WikiEditorsInsightActionBarProps = {
+  handleSearchKeyword: (e: ChangeEvent<HTMLInputElement>) => void
+  handleSortChange: () => void
+  isOpenFilter: boolean
+  onCloseFilter: () => void
+  ApplyFilterItems: (e: React.FormEvent<HTMLFormElement>) => void
+  onToggleFilter: () => void
+  FilterArray: {
+    id: string
+    value: string
+  }[]
+  sortIcon: JSX.Element
+  setChecked: (checked: number) => void
+  checked: number
+  setFilterEditors: (editors: string[]) => void
+  setPaginateOffset: (offset: number) => void
+}
+
+export type PromoteCreatedWikisModalProps = {
+  isOpen: boolean
+  onClose: () => void
+  wikiChosenTitle: string
+  wikiChosenId: string
+  hideFunc: () => void
+}
+
+export type ContentProps = {
+  activeStep: number
+  steps?: {
+    label: string
+    description: string
+  }[]
+  HompageSelected?: () => void
+  buttonOne?: string
+  buttonTwo?: string
+  step2Titles: string
+  promotedWikis: CreatedWikisCount[] | undefined
+  Data: CreatedWikisCount | undefined
+  setValue: (value: React.SetStateAction<string>) => void
+  TrendingwikiSelected?: () => Promise<void>
+}
+
+export type WikisTableProps = {
+  wikiTableData: Wikis[]
+  findSection: (promotedNum: number) => void
+  shouldPromote: (wikiTitle: string, id: string) => void
+  shouldArchive: (ishidden: boolean, wikiId: string) => void
+}
+
+export type WikiCreatedFooterProps = {
+  activatePrevious: boolean
+  scrolltoTableTop: () => void
+  setPaginateOffset: React.Dispatch<React.SetStateAction<number>>
+  paginateOffset: number
+  setActivatePrevious: (value: React.SetStateAction<boolean>) => void
+  wikis: CreatedWikisCount[] | undefined
+  nextBtnDisabled: boolean
+}
+
+export type WikiCreatedActionBarProps = {
+  setsearchKeyWord: (value: React.SetStateAction<string>) => void
+  sortIcon: JSX.Element
+  isOpen: boolean
+  onClose: () => void
+  onToggle: () => void
+  handleSortChange: (value: number) => void
+  checked: number
+  setChecked: (value: React.SetStateAction<number>) => void
+  setPaginateOffset: (value: React.SetStateAction<number>) => void
+  setFilterItems: (value: React.SetStateAction<unknown[] | undefined>) => void
 }
